--- app/ClientLayout.tsx
@@ -1,29 +1,0 @@
-"use client"
-
-import type React from "react"
-import { Inter } from "next/font/google"
-import { ThemeProvider } from "@/components/theme-provider"
-import { Toaster } from "@/components/ui/toaster"
-import { AuthProvider } from "@/lib/AuthContext"
-import "./globals.css"
-
-const inter = Inter({ subsets: ["latin"] })
-
-interface ClientLayoutProps {
-  children: React.ReactNode
-}
-
-export default function ClientLayout({ children }: ClientLayoutProps) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={inter.className}>
-        <ThemeProvider defaultTheme="dark">
-          <AuthProvider>
-            {children}
-            <Toaster />
-          </AuthProvider>
-        </ThemeProvider>
-      </body>
-    </html>
-  )
-}
