+++ app/auth/confirm/route.ts
@@ -1,7 +1,7 @@
 import { type EmailOtpType } from '@supabase/supabase-js'
 import { type NextRequest } from 'next/server'
 
-import { createClient } from '@/lib/supabase/server'
+import { createClient } from '@/lib/modassembly/supabase/server'
 import { redirect } from 'next/navigation'
 
 export async function GET(request: NextRequest) {
