+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type React from "react"
 import { Inter } from "next/font/google"
 import { ThemeProvider } from "@/components/theme-provider"
 import { Toaster } from "@/components/ui/toaster"
-import { createClient } from "@/lib/supabase/server"
+import { createClient } from "@/lib/modassembly/supabase/server"
 import { headers } from "next/headers"
 import { redirect } from "next/navigation"
 import "./globals.css"
