+++ lib/modassembly/supabase/auth/actions.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-import { createClient } from '@/lib/supabase/server'
+import { createClient } from '@/lib/modassembly/supabase/server'
 
 type ActionResult = {
   error?: string;
