+++ lib/modassembly/supabase/database/seats.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/lib/supabase/client";
+import { createClient } from "@/lib/modassembly/supabase/client";
 
 /**
  * Fetch a seat ID based on table and seat label
