+++ lib/modassembly/supabase/database/suggestions.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/lib/supabase/client'
+import { createClient } from '@/lib/modassembly/supabase/client'
 
 // Type definitions
 type OrderSuggestion = {
