--- scripts/create_tables_seats.sql
@@ -1,1 +1,0 @@
-
