--- supabase/.temp/cli-latest
@@ -1,1 +1,0 @@
-v2.22.12
