--- supabase/.temp/gotrue-version
@@ -1,1 +1,0 @@
-v2.170.0
