--- supabase/.temp/pooler-url
@@ -1,1 +1,0 @@
-postgresql://postgres.eiipozoogrrfudhjoqms:[YOUR-PASSWORD]@aws-0-us-east-1.pooler.supabase.com:6543/postgres
