--- supabase/.temp/postgres-version
@@ -1,1 +1,0 @@
-15.8.1.054
