--- supabase/.temp/rest-version
@@ -1,1 +1,0 @@
-v12.2.3
